cli: remove all traces of old locking mechanism from the CLI
At the same, time put in place a mechanism that enables the patch management
interface to be determined at run time.  The aim here is to facilitate
modifying the layout of the database directory etc. in ways which are
incompatible with older playgrounds but still be runnable on older playgrounds.
I.e. if we detect that we're in an older playground we load the appropriate
interface.

Signed-off-by: Peter Williams <pwil3058@gmail.com>