patch_db: use dict proxy rather than wrapper for file data
All direct access to pickled data now occurs via __getitem__/__setitem__
making it easy to see where such access occurs. Plus the basic mechanisms
needed for ensuring the pickled data is consistent is now in place so this
functionality can be added to the "validate" command used in the testing
suite.

Signed-off-by: Peter Williams <pwil3058@gmail.com>