push: don't modify ecode returned by the application code
We're no longer pretending to mimic patch. A return value of 1 now means that
a refresh is required as the application was not pure (i.e. the file being
patched was not an exact match to a diff's expectations). Also means that a
prudent programmer would check the effected files to make sure the applied
changes are correct.

Signed-off-by: Peter Williams <pwil3058@gmail.com>